Update the datatype of phone number column

+++ create-table-scripts.sql
@@ -1,21 +1,21 @@
 USE social_media_db;
 CREATE TABLE USER_PROFILE(
-	id int NOT NULL,
+    id int NOT NULL,
     profile_picture_uri varchar(255) NOT NULL,
     name varchar(255) NOT NULL,
     email varchar(255) NOT NULL,
-    phone_number int NOT NULL,
+    phone_number BIGINT NOT NULL,
     PRIMARY KEY(id)
 );
 
 CREATE TABLE SKILLS(
-	id int NOT NULL,
+    id int NOT NULL,
     name varchar(255) NOT NULL,
     PRIMARY KEY(id)
 );
 
 CREATE TABLE USER_SKILLS(
-	id int NOT NULL,
+    id int NOT NULL,
     user_id int NOT NULL,
     skill_id int NOT null,
     PRIMARY KEY(id),
